解決Merge後衝突問題。

+++ SuFood專案MVC/SuFood/Models/Coupon.cs
@@ -16,7 +16,7 @@ namespace SuFood.Models
         public int CouponId { get; set; }
         public string CouponDescription { get; set; }
         public string CouponName { get; set; }
-        public int? CouponMinusCost { get; set; }
+        public decimal? CouponMinusCost { get; set; }
         public int? MinimumPurchasingAmount { get; set; }
         public DateTime? CouponStartDate { get; set; }
         public DateTime? CouponEndDate { get; set; }
